+++ packages/graph/vite.config.mjs
@@ -1,8 +1,6 @@
 import { createConfig } from '@warp-drive/internal-config/vite/config.js';
 
-export const externals = [
-  '@ember/debug', // assert, deprecate
-];
+export const externals = [];
 
 export const entryPoints = ['./src/-private.ts'];
