+++ warp-drive-packages/core/src/graph/-private/edges/collection.ts
@@ -1,11 +1,10 @@
-import type { StableRecordIdentifier } from '@warp-drive/core-types';
-import type { CollectionRelationship } from '@warp-drive/core-types/cache/relationship';
-import type { Links, Meta, PaginationLinks } from '@warp-drive/core-types/spec/json-api-raw';
-
-import { computeLocalState } from '../-diff';
-import type { UpgradedMeta } from '../-edge-definition';
-import type { RelationshipState } from '../-state';
-import { createState } from '../-state';
+import type { CollectionRelationship } from '../../../types/cache/relationship.ts';
+import type { StableRecordIdentifier } from '../../../types/identifier.ts';
+import type { Links, Meta, PaginationLinks } from '../../../types/spec/json-api-raw.ts';
+import { computeLocalState } from '../-diff.ts';
+import type { UpgradedMeta } from '../-edge-definition.ts';
+import type { RelationshipState } from '../-state.ts';
+import { createState } from '../-state.ts';
 
 export interface CollectionEdge {
   definition: UpgradedMeta;
