+++ warp-drive-packages/core/src/graph/-private/operations/update-relationship.ts
@@ -1,18 +1,18 @@
 import { warn } from '@ember/debug';
 
-import type Store from '@ember-data/store';
 import { assert } from '@warp-drive/build-config/macros';
-import type { StableRecordIdentifier } from '@warp-drive/core-types';
-import type { UpdateResourceRelationshipOperation } from '@warp-drive/core-types/cache/operations';
-import type { UpdateRelationshipOperation } from '@warp-drive/core-types/graph';
+
+import type { Store } from '../../../index.ts';
+import type { StableRecordIdentifier } from '../../../types.ts';
+import type { UpdateResourceRelationshipOperation } from '../../../types/cache/operations.ts';
+import type { UpdateRelationshipOperation } from '../../../types/graph.ts';
 import type {
   ExistingResourceIdentifierObject,
   NewResourceIdentifierObject,
-} from '@warp-drive/core-types/spec/json-api-raw';
-
-import { isBelongsTo, isHasMany, notifyChange } from '../-utils';
-import type { Graph } from '../graph';
-import _normalizeLink from '../normalize-link';
+} from '../../../types/spec/json-api-raw.ts';
+import { isBelongsTo, isHasMany, notifyChange } from '../-utils.ts';
+import type { Graph } from '../graph.ts';
+import _normalizeLink from '../normalize-link.ts';
 
 type IdentifierCache = Store['identifierCache'];
